+++ tests/di/Fake/FakeAssistedDbModule.php
@@ -9,5 +9,8 @@ class FakeAssistedDbModule extends AbstractModule
     protected function configure()
     {
         $this->bind(FakeAbstractDb::class)->toProvider(FakeAssistedDbProvider::class);
+        $this->bind(FakeAssistedParamsConsumer::class);
+        $this->bind(FakeAssistedDbProvider::class);
+        $this->bind(FakeAssistedInjectDb::class);
     }
 }