+++ tests/di/Fake/FakeBcParameterQualifierModule.php
@@ -10,6 +10,7 @@ class FakeBcParameterQualifierModule extends AbstractModule
 {
     protected function configure(): void
     {
+        $this->bind(FakeClassWithBcParameterQualifier::class);
         // Bind with FakeInjectOne qualifier
         $this->bind(FakeGearStickInterface::class)
             ->annotatedWith(FakeInjectOne::class)