+++ tests/di/Fake/FakeConstantModule.php
@@ -8,6 +8,7 @@ class FakeConstantModule extends AbstractModule
 {
     protected function configure()
     {
+        $this->bind(FakeConstantConsumer::class);
         $this->bind()->annotatedWith(FakeConstant::class)->toInstance('kuma');
     }
 }