+++ tests/di/Fake/FakeInternalTypeModule.php
@@ -12,6 +12,7 @@ class FakeInternalTypeModule extends AbstractModule
 {
     protected function configure()
     {
+        $this->bind(FakeInternalTypes::class);
         $this->bind('')->annotatedWith('type-bool')->toInstance(false);
         $this->bind('')->annotatedWith('type-int')->toInstance(1);
         $this->bind('')->annotatedWith('type-string')->toInstance('1');