+++ tests/di/Fake/FakeWalkRobotOtherDepModule.php
@@ -8,6 +8,7 @@ class FakeWalkRobotOtherDepModule extends AbstractModule
 {
     protected function configure()
     {
+        $this->bind(FakeWalkRobot::class);
         $this->bind(FakeRobotInterface::class)->to(FakeRobot::class);
         $this->bind(FakeLegInterface::class)->toProvider(FakeWalkRobotLegProviderWithOtherDep::class);
     }