+++ tests/di/MultiBinding/MultiBindingModuleTest.php
@@ -40,6 +40,14 @@ class MultiBindingModuleTest extends TestCase
         $this->module = new class extends AbstractModule {
             protected function configure(): void
             {
+                $this->bind(FakeMultiBindingConsumer::class);
+                $this->bind(FakeMultiBindingAnnotation::class);
+                $this->bind(FakeEngine::class);
+                $this->bind(FakeEngine2::class);
+                $this->bind(FakeEngine3::class);
+                $this->bind(FakeRobot::class);
+                $this->bind(FakeRobotProvider::class);
+                $this->bind(FakeSetNotFoundWithMap::class);
                 $engineBinder = MultiBinder::newInstance($this, FakeEngineInterface::class);
                 $engineBinder->addBinding('one')->to(FakeEngine::class);
                 $engineBinder->addBinding('two')->to(FakeEngine2::class);
@@ -206,7 +214,12 @@ class MultiBindingModuleTest extends TestCase
     public function testSetNotFoundInProvider(): void
     {
         $this->expectException(SetNotFound::class);
-        $injector = new Injector();
+        $injector = new Injector(new class extends AbstractModule {
+            protected function configure(): void
+            {
+                $this->bind(FakeSetNotFoundWithProvider::class);
+            }
+        });
         $injector->getInstance(FakeSetNotFoundWithProvider::class);
     }
 
@@ -220,6 +233,7 @@ class MultiBindingModuleTest extends TestCase
         $module = new class extends AbstractModule {
             protected function configure(): void
             {
+                $this->bind(FakeMultiBindingConsumer::class);
                 MultiBinder::newInstance($this, FakeEngineInterface::class);
                 MultiBinder::newInstance($this, FakeRobotInterface::class);
             }
@@ -240,7 +254,12 @@ class MultiBindingModuleTest extends TestCase
      */
     public function testSetNotBound(): void
     {
-        $injector = new Injector(new NullModule());
+        $injector = new Injector(new class (new NullModule()) extends AbstractModule {
+            protected function configure(): void
+            {
+                $this->bind(FakeMultiBindingConsumer::class);
+            }
+        });
         try {
             $injector->getInstance(FakeMultiBindingConsumer::class);
             self::fail('SetNotBound must be thrown when no MultiBinder declared the #[Set] interface.');
@@ -265,6 +284,7 @@ class MultiBindingModuleTest extends TestCase
         $module = new class extends AbstractModule {
             protected function configure(): void
             {
+                $this->bind(FakeMultiBindingConsumer::class);
                 MultiBinder::newInstance($this, FakeRobotInterface::class)->addBinding('robot')->to(FakeRobot::class);
                 $engineBinder = MultiBinder::newInstance($this, FakeEngineInterface::class);
                 $engineBinder->addBinding('one')->to(FakeEngine::class);